Added launch image

+++ RWTFlickrSearch/View/RWTFlickrFilterViewController/RWTFlickrFilterViewController.m
@@ -68,7 +68,7 @@
     self.sectionTextField.inputView = self.sectionPickerView;
     
     self.sectionTextField.text = [self.viewModel.selectedFilterOptions.selectedSection prettyName];
-    int sectionPickerIndexSelected = [[self.viewModel getArrayOfAllSectionTypes] indexOfObject:self.viewModel.selectedFilterOptions.selectedSection];
+    long sectionPickerIndexSelected = [[self.viewModel getArrayOfAllSectionTypes] indexOfObject:self.viewModel.selectedFilterOptions.selectedSection];
     self.viewModel.lastSectionIndexSelected = sectionPickerIndexSelected;
     [self.sectionPickerView selectRow:sectionPickerIndexSelected inComponent:0 animated:NO];
 }
@@ -81,7 +81,7 @@
     self.viewTypeTextField.inputView = self.viewTypePickerView;
     
     self.viewTypeTextField.text = [self.viewModel.selectedViewType prettyName];
-    int viewPickerIndexSelected = [[self.viewModel getArrayOfAllViewTypes] indexOfObject:self.viewModel.selectedViewType];
+    long viewPickerIndexSelected = [[self.viewModel getArrayOfAllViewTypes] indexOfObject:self.viewModel.selectedViewType];
     [self.viewTypePickerView selectRow:viewPickerIndexSelected inComponent:0 animated:NO];
 
 }
@@ -102,7 +102,7 @@
     self.sortTextField.inputView = self.sortPickerView;
     
     self.sortTextField.text = [self.viewModel.selectedFilterOptions.selectedSort prettyName];
-    int sortPickerIndexSelected = [[self.viewModel getArrayOfAllSortTypes:[self userHasSelectedUserSection]] indexOfObject:self.viewModel.selectedFilterOptions.selectedSort];
+    long sortPickerIndexSelected = [[self.viewModel getArrayOfAllSortTypes:[self userHasSelectedUserSection]] indexOfObject:self.viewModel.selectedFilterOptions.selectedSort];
     [self.sortPickerView selectRow:sortPickerIndexSelected inComponent:0 animated:NO];
 }
 
@@ -115,7 +115,7 @@
     
     if (self.viewModel.selectedWindow) {
         self.windowTextField.text = [self.viewModel.selectedWindow prettyName];
-        int windowPickerIndexSelected = [[self.viewModel getArrayOfAllWindowTypes] indexOfObject:self.viewModel.selectedWindow];
+        long windowPickerIndexSelected = [[self.viewModel getArrayOfAllWindowTypes] indexOfObject:self.viewModel.selectedWindow];
         self.viewModel.lastWindowTypeIndexSelected = windowPickerIndexSelected;
         [self.windowPickerView selectRow:windowPickerIndexSelected inComponent:0 animated:NO];
         
@@ -206,7 +206,7 @@
 
 - (BOOL)userHasSelectedUserSection{
     
-    int indexSelected = self.viewModel.lastSectionIndexSelected;
+    long indexSelected = self.viewModel.lastSectionIndexSelected;
     
     RWTImgurSection *sectionSelected = [self.viewModel getArrayOfAllSectionTypes][indexSelected];
